+++ client/components/App.jsx
@@ -1,32 +1,16 @@
 import React from "react"
 import Header from "./Header"
-import Form from "./Form"
-import Result from "./Result"
+import Theory from "./Theory"
 import Footer from "./Footer"
 
-class App extends React.Component {
-constructor(props) {
-  super(props)
-
-  this.state = { formName: "" }
-}
-
-callbackFormText = (dataFromForm) => {
-  if (dataFromForm !== undefined) {
-    this.setState({formName: dataFromForm})
-  }
-}
-
-render() {
+const App = () => {
   return (
     <div>
       <Header />
-      <Result resultText={this.state.formName}/>
-      <Form getForm={this.callbackFormText}/>
+      <Theory />
       <Footer />
     </div>
   )
-}
 }
 
 export default App
